Refactor menuitem views
